requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.11.0=pypi_0
anaconda-clean=1.1.0=py38_1
appnope=0.1.0=py38_1001
argon2-cffi=20.1.0=py38haf1e3a3_1
astunparse=1.6.3=pypi_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
brotlipy=0.7.0=py38h9ed2024_1003
ca-certificates=2020.10.14=0
cachetools=4.1.1=pypi_0
catalogue=1.0.0=py38h50d1736_2
certifi=2020.11.8=py38hecd8cb5_0
cffi=1.14.3=py38h2125817_2
chardet=3.0.4=py38hecd8cb5_1003
click=7.1.2=py_0
cryptography=3.2.1=py38hbcfaee0_1
cycler=0.10.0=py38_0
cymem=2.0.4=py38hf416894_1
cython-blis=0.7.3=py38hfb243c8_0
de-core-news-sm=2.3.0=pypi_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
en-core-web-md=2.3.1=pypi_0
en-core-web-sm=2.3.1=pypi_0
entrypoints=0.3=py38_0
freetype=2.10.4=ha233b18_0
gast=0.3.3=pypi_0
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.33.2=pypi_0
h5py=2.10.0=pypi_0
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2019.4=233
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38h01d92e1_0
ipython_genutils=0.2.0=py38_0
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=he5867d9_2
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py38hecd8cb5_0
jupyterlab=2.2.6=py_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=1.2.0=py_0
keras=2.4.3=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.0=py38h23ab428_0
langdetect=1.0.8=pyh9f0ad1d_0
lcms2=2.11=h92f6f08_0
libcxx=11.0.0=h4c3b8ed_1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
llvm-openmp=11.0.0=h73239a0_1
lz4-c=1.9.2=h79c402e_3
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py38h1de35cc_1
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h181983e_0
mistune=0.8.4=py38h1de35cc_1001
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
msgpack-python=1.0.0=py38h04f5b5a_1
murmurhash=1.0.4=py38hf416894_2
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
nltk=3.5=py_0
notebook=6.1.4=py38_0
numpy=1.18.5=pypi_0
numpy-base=1.19.2=py38hcfb5961_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openssl=1.1.1h=haf1e3a3_0
opt-einsum=3.3.0=pypi_0
packaging=20.4=py_0
pandas=1.1.3=py38hb1e8313_0
pandoc=2.11=h0dc7051_0
pandocfilters=1.4.3=py38hecd8cb5_1
parso=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=py38_1001
pillow=8.0.1=py38h5270095_0
pip=20.2.4=py38hecd8cb5_0
plac=0.9.6=py38_1
preshed=3.0.4=py38hf416894_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
protobuf=3.14.0=pypi_0
ptyprocess=0.6.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pygments=2.7.2=pyhd3eb1b0_0
pyopenssl=19.1.0=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pypdf2=1.26.0=py_2
pyrsistent=0.17.3=py38haf1e3a3_0
pysocks=1.7.1=py38_1
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=py_0
python-dotenv=0.15.0=pyhd8ed1ab_0
python_abi=3.8=1_cp38
pytz=2020.1=py_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=py38hb1e8313_1
readline=8.0=h1de35cc_0
regex=2020.10.15=py38haf1e3a3_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-learn=0.23.2=py38h959d312_0
scipy=1.5.2=py38h2515648_0
send2trash=1.5.0=py38_0
setuptools=50.3.1=py38hecd8cb5_1
six=1.15.0=py38hecd8cb5_0
spacy=2.3.4=py38hd731e05_0
sqlite=3.33.0=hffcf06c_0
srsly=1.0.4=py38hf416894_0
tensorboard=2.4.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow=2.3.1=pypi_0
tensorflow-estimator=2.3.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.1=py38_0
testpath=0.4.4=py_0
thinc=7.4.3=py38hd731e05_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
tornado=6.0.4=py38h1de35cc_1
tqdm=4.53.0=pyhd8ed1ab_1
traitlets=5.0.5=py_0
urllib3=1.25.11=py_0
wasabi=0.8.0=pyh9f0ad1d_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pypi_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.3=hb1e8313_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0

spacy~=2.3.4
langdetect~=1.0.8
python-dotenv~=0.15.0
pandas~=1.1.3
scikit-learn~=0.23.2
